+++ doc/conf.py
@@ -31,6 +31,7 @@ master_doc = "index"
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = [
+    "sphinxcontrib.bibtex",
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
@@ -39,11 +40,14 @@ extensions = [
     "sphinx.ext.linkcode",
     "sphinx_autodoc_typehints",
     "sphinx_copybutton",
+    "nbsphinx",
 ]
 
 # Create links to references within kikuchipy's documentation to these packages.
 intersphinx_mapping = {
     "dask": ("https://docs.dask.org/en/latest", None),
+    "diffpy.structure": ("https://www.diffpy.org/diffpy.structure", None),
+    "diffsims": ("https://diffsims.readthedocs.io/en/latest", None),
     "hyperspy": ("http://hyperspy.org/hyperspy-doc/current", None),
     "matplotlib": ("https://matplotlib.org", None),
     "numpy": ("https://docs.scipy.org/doc/numpy", None),
@@ -64,7 +68,8 @@ templates_path = [
 # directories to ignore when looking for source files. This image also affects
 # html_static_path and html_extra_path.
 exclude_patterns = [
-    "_build",
+    "build",
+    "_static/v0.2.0/*.ipynb",
 ]
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for a
@@ -97,6 +102,40 @@ html_theme_options = {
 
 # Figure references
 numfig = True
+
+# nbsphinx configuration
+# Taken from nbsphinx' own nbsphinx configuration file.
+# This is processed by Jinja2 and inserted before each notebook
+nbsphinx_prolog = r"""
+{% set docname = 'doc/' + env.doc2path(env.docname, base=None) %}
+
+.. raw:: html
+
+    <style>a:hover { text-decoration: underline; }</style>
+
+    <div class="admonition note">
+      This page was generated from
+      <a class="reference external" href="https://github.com/pyxem/kikuchipy/blob/{{ env.config.release|e }}/{{ docname|e }}">{{ docname|e }}</a>.
+      <script>
+        if (document.location.host) {
+          $(document.currentScript).replaceWith(
+            '<a class="reference external" ' +
+            'href="https://nbviewer.jupyter.org/url' +
+            (window.location.protocol == 'https:' ? 's/' : '/') +
+            window.location.host +
+            window.location.pathname.slice(0, -4) +
+            'ipynb">View in <em>nbviewer</em></a>.'
+          );
+        }
+      </script>
+    </div>
+
+.. raw:: latex
+
+    \nbsphinxstartnotebook{\scriptsize\noindent\strut
+    \textcolor{gray}{The following section was generated from
+    \sphinxcode{\sphinxupquote{\strut {{ docname | escape_latex }}}} \dotfill}}
+"""
 
 
 def linkcode_resolve(domain, info):