+++ kikuchipy/draw/colors.py
@@ -16,16 +16,32 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Simulations returned by a generator and handling of Kikuchi bands and
-zone axes.
-"""
+"""Color palettes for plotting Kikuchi bands."""
 
-from kikuchipy.simulations.geometrical_ebsd_simulation import (
-    GeometricalEBSDSimulation,
-)
-from kikuchipy.simulations import features
-
-__all__ = [
-    "GeometricalEBSDSimulation",
-    "features",
+# Kikuchi band color palette from EDAX TSL Data Collection version 7
+RED = (1, 0, 0)
+GREEN = (0, 1, 0)
+BLUE = (0, 0, 1)
+YELLOW = (1, 1, 0)
+CYAN = (0, 1, 1)
+PINK = (1, 0, 1)
+BROWN = (0.5, 0, 0)
+EARTH = (0.5, 0.5, 0)
+PURPLE = (0.5, 0, 0.5)
+DARK_GREEN = (0, 0.5, 0)
+DARKER_GREEN = (0, 0.5, 0.5)
+DARK_BLUE = (0, 0, 0.5)
+TSL_COLORS = [
+    RED,
+    YELLOW,
+    GREEN,
+    BLUE,
+    PINK,
+    CYAN,
+    BROWN,
+    DARK_GREEN,
+    PURPLE,
+    DARK_BLUE,
+    EARTH,
+    DARKER_GREEN,
 ]