+++ kikuchipy/projections/ebsd_projections.py
@@ -16,6 +16,10 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
+"""Rotations to align the EBSD detector with the tilted sample. Notation
+from [Britton2016]_.
+"""
+
 from diffpy.structure import Lattice
 import numpy as np
 from orix.quaternion.rotation import Rotation