+++ kikuchipy/projections/lambert_projection.py
@@ -96,9 +96,7 @@ class LambertProjection:
             abs_yx, c_x * np.sin(true_term), c_y * np.cos(false_term)
         )
         cart[..., 2] = np.where(
-            abs_yx,
-            1 - (2 * (X ** 2)) / np.pi,
-            1 - (2 * (Y ** 2)) / np.pi,
+            abs_yx, 1 - (2 * (X ** 2)) / np.pi, 1 - (2 * (Y ** 2)) / np.pi,
         )
 
         return Vector3d(cart)